+++ armory/art_experimental/__init__.py
@@ -2,3 +2,4 @@
 This subpackage will contain experimental ART features that are not merged into
 the PyPi package.
 """
+from armory import delayed_imports
