+++ armory/baseline_models/__init__.py
@@ -1,3 +1,4 @@
 """
 Baseline models for use in the ARMORY system.
 """
+from armory import delayed_imports
