+++ armory/data/__init__.py
@@ -1,3 +1,4 @@
 """
 Data methods in ARMORY
 """
+from armory import delayed_imports
