+++ armory/scenarios/__init__.py
@@ -1,3 +1,4 @@
 """
 Evaluation scenarios for novel defenses
 """
+from armory import delayed_imports
